sso注销用户删除企业成员记录

+++ src/main/java/com/neo/service/clear/impl/ClearService.java
@@ -1,5 +1,6 @@
 package com.neo.service.clear.impl;
 
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -7,6 +8,8 @@ import com.neo.commons.cons.DefaultResult;
 import com.neo.commons.cons.EnumResultCode;
 import com.neo.commons.cons.IResult;
 import com.neo.commons.properties.PtsProperty;
+import com.neo.model.dto.UserClearDto;
+import com.neo.model.dto.UserClearRequestDto;
 import com.neo.model.po.PtsAuthPO;
 import com.neo.model.qo.FcsFileInfoQO;
 import com.neo.service.auth.IAuthService;
@@ -53,17 +56,31 @@ public class ClearService implements IClearService{
      * @return
      */
     @Override
-    public IResult<Integer> clearUserData(String id) {
-    	Long userId = Long.valueOf(id);
-    	FcsFileInfoQO fcsFileInfoQO = new FcsFileInfoQO();
-    	fcsFileInfoQO.setUserID(userId);
-    	
-    	PtsAuthPO ptsAuthPO =new PtsAuthPO();
-    	ptsAuthPO.setUserid(userId);
-    	
-    	ptsConvertService.deletePtsConvert(fcsFileInfoQO);
-    	iAuthService.deletePtsAuth(ptsAuthPO);
-    	
+    public IResult<Integer> clearUserData(UserClearRequestDto userClearRequestDto) {
+    	if(StringUtils.isNotBlank(userClearRequestDto.getId())) {
+    		Long userId = Long.valueOf(userClearRequestDto.getId());
+        	FcsFileInfoQO fcsFileInfoQO = new FcsFileInfoQO();
+        	fcsFileInfoQO.setUserID(userId);
+        	
+        	PtsAuthPO ptsAuthPO =new PtsAuthPO();
+        	ptsAuthPO.setUserid(userId);
+        	
+        	ptsConvertService.deletePtsConvert(fcsFileInfoQO);
+        	iAuthService.deletePtsAuth(ptsAuthPO);
+        	
+        	//删企业成员记录
+        	if(userClearRequestDto.getMembers().length > 0) {
+        		for(UserClearDto userClearDto : userClearRequestDto.getMembers()) {
+        			if(StringUtils.isNotBlank(userClearRequestDto.getId())) {
+        				userId = Long.valueOf(userClearDto.getId());
+            			fcsFileInfoQO.setUserID(userId);
+            			ptsAuthPO.setUserid(userId);
+            			ptsConvertService.deletePtsConvert(fcsFileInfoQO);
+                    	iAuthService.deletePtsAuth(ptsAuthPO);
+        			}
+        		}
+        	}
+    	}
     	return DefaultResult.successResult();
     }
